requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
matplotlib
Pillow
